edge/arducam/requirements.txt
# Arducam Pivistation 5 (IMX500) Dependencies
# For use with Raspberry Pi OS and picamera2

# Camera and AI inference
picamera2>=0.3.12

# Core scientific computing
numpy>=1.24.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# MQTT communication
paho-mqtt>=1.6.1

# Configuration management
PyYAML>=6.0

# LoRa communication (using SX127x modules)
# Note: pyLoRa is a placeholder - actual implementation may use RPi.GPIO + spidev
pyLoRa>=0.3.1

# Data structures
dataclasses>=0.6; python_version < '3.7'

# Logging and utilities
python-dateutil>=2.8.2
